# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateReviewRequest:
    properties:
      comment:
        type: string
      job_id:
        type: integer
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - job_id
    - rating
    type: object
...
  models.UserRatingStats:
    properties:
      average_rating:
        type: number
      five_stars:
        type: integer
      four_stars:
        type: integer
      one_star:
        type: integer
      three_stars:
        type: integer
      total_reviews:
        type: integer
      two_stars:
        type: integer
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API для приложения MoveShare
  title: MoveShare API
  version: "1.0"
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Creates a new review for a completed job
      parameters:
      - description: Review creation data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new review
      tags:
      - Reviews
  /reviews/average/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves the average rating and total review count for a specific
        user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User average rating and review count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user average rating
      tags:
      - Reviews
  /reviews/stats/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed rating statistics for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User rating statistics
          schema:
            additionalProperties:
              $ref: '#/definitions/models.UserRatingStats'
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user rating statistics
      tags:
      - Reviews
  /reviews/user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves reviews for a specific user with pagination
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User reviews with pagination
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user reviews
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in format: Bearer {your_token_here}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"